+++ test/unit/test_repos.py
@@ -1,5 +1,10 @@
 import subprocess
-from unittest.mock import patch
+from unittest.mock import (
+    MagicMock,
+    patch,
+)
+
+from github import Repository
 
 from github_archive import GithubArchive
 from github_archive.archive import (
@@ -8,7 +13,9 @@ from github_archive.archive import (
 )
 from github_archive.repos import (
     _archive_repo,
+    _fork_repo,
     iterate_repos_to_archive,
+    iterate_repos_to_fork,
     view_repos,
 )
 
@@ -136,4 +143,34 @@ def test_archive_repo_called_process_error(mock_logger, mock_subprocess, mock_gi
     github_archive = GithubArchive()
     _archive_repo(github_archive, mock_git_asset, 'github_archive', operation)
 
+    mock_logger.assert_called_once()
+
+
+@patch('github_archive.repos._fork_repo')
+def test_iterate_repos_to_fork(mock_fork_repo):
+    repo = MagicMock(spec=Repository.Repository)
+    iterate_repos_to_fork([repo])
+
+    mock_fork_repo.assert_called_once()
+
+
+@patch('logging.Logger.info')
+@patch('github.Repository.Repository.create_fork')
+def test_fork_repo_success(mock_create_fork, mock_logger):
+    repo = MagicMock(spec=Repository.Repository)
+    repo.create_fork = mock_create_fork
+    _fork_repo(repo)
+
+    mock_create_fork.assert_called_once()
+    mock_logger.assert_called_once()
+
+
+@patch('logging.Logger.error')
+@patch('github.Repository.Repository.create_fork', side_effect=Exception())
+def test_fork_repo_failure(mock_create_fork, mock_logger):
+    repo = MagicMock(spec=Repository.Repository)
+    repo.create_fork = mock_create_fork
+    _fork_repo(repo)
+
+    mock_create_fork.assert_called_once()
     mock_logger.assert_called_once()